+++ internal/image/imageos/imageos.go
@@ -464,6 +464,14 @@ func (imageOs *ImageOs) initDebLocalRepoWithinInstallRoot(installRoot string) er
 		return fmt.Errorf("failed to get chroot environment path for install root %s: %w", installRoot, err)
 	}
 
+	if err := imageOs.chrootEnv.UpdateChrootLocalRepoMetadata(
+		chroot.ChrootRepoDir,
+		imageOs.template.Target.Arch,
+		false,
+	); err != nil {
+		return fmt.Errorf("failed to refresh local debian repository metadata: %w", err)
+	}
+
 	// from local.list
 	repoPath := filepath.Join(chrootInstallRoot, "/cdrom/cache-repo")
 	chrootPkgCacheDir := imageOs.chrootEnv.GetChrootPkgCacheDir()